chore: CI/CD Implementation
The following yml file, is meant to harden the status of the development
branch, since it builds and runs the test online, it will fail if any of
the tasks is not completed successfuly.

The first task: Build, only builds the project avoiding the tests
execution so that it only focuses in the build of the project.

The next task: Test, only runs the tests of the backend. So that the
code works as intended.