fix: erase repeated elements from generation pools
The generation pools(first_name_pool, first_last_name_pool,
second_last_name_pool) had repeated elements because of that the
generator was not working as expected.

After deleting the repeated elements, I have added a test to check the
integrity of the tree pools so that we can discard this kind of
anomally.