chore: name generation testing.
Testing that the names generation implementation makes the an expected
number of names according to the requirementes.

The generation stacks: first_name_poool, first_last_name_pool and
second_last_name_pool are being tested for:
- Uniqueness
- Repetition Threshold

But the second_name is not being tested because it is not needed because
if the other pass it means that this one will as well.