chore: isolate check tasks to a job
Implemented a job to run especically the check tasks while the other
jobs avoid or ignore this task.

I have changed the scope of the checkstyle revisions, now the checks
will be run only for the code in src/main hence the :checkStyleMain and
that's why the :checkStyleTest task is being ignored.