fix: add gradle wrapper jar file
The .gitignore file from the eureka folder was configured to ignore the
gradle folder that's where the jar of the wrapper is. So I have changed
that configuration to allow the tracking of the gradle folder.

While adding the jar file I noticed it was using a different version
than from the other services that's why I have updated it to the general
version of 8.3