chore: change gradle plugin
The build.gradle of each service was configured with the java plugin,
this does not provide distribution so each service would have to be
build in each machine that we wanted it to run. This new plugin
"application" creates a distribution tar loacated in the
build/distributions path, that's the file we intent to distribute to
each machine or node.