fix: run_services log flag
This command: `./scripts/fake_orgasm --run --log scripts` is supposed to
log the ouput of the services in the log folder, this logging behaviour
is done by the run_services script. What was happening was that when
using this flag the contents of the logs folder where not being updated
as expected.

Solved by changing the argument number where the getops starts checking.
Initially it was checking from $1 but now it does from $3 in that way
the first two arguments are: $1 for the execution order file,
and $2 the directory where the bins are. This correction has succesfully
fixed the issue.