refactor: run services redesign
The run_services script has undergone a significant redesign
to offer more control and flexibility to the user. Instead
of automatically searching for binaries, it now expects the
binary as an argument. This change was necessary because each
service must be executed in a specific order, and previous
attempts to start them simultaneously led to synchronization
issues. This approach lacked testability, which prompted the refactoring.

The previous implementation heavily relied on the fake_orgasm
script as the main orchestrator, driven by the
service_execution_order file. However, in the new approach
where the user has more decision-making power, it made sense
to remove this script.

Additionally, a new utility script called "utils" has been
introduced. It allows users to check whether a port is in
use and provides the capability to terminate services using
the TERM signal.

Note: Utils should kill the process in an specific order, leaving
gateway and eureka at the end which is currently not supported.