chore: update money exchange CI/CD workflow
I have updated the workflow to run test scripts and pack the binaries
and logs. In this way we can know if the service is running as expected
and can recieve and send requests.

The test scripts that will be executed by the CI/CD pipeline for
the services will be located in the "services_testing" folder.

Created two scripts to test the money_exchange microservice.
- service_startup: it creates the enviroment to run the service tests
- simple_test: this script runs a simple query to check if the gateway
  is connected.