fix: add specific paths to workflows
To avoid triggering every workflow on every push, I have added
the respective path to each workflow so that it only triggers a
workflow for a specific service. This behavior applies only when
the pushed commit is part of a pull request. When the push is
made to the "develop" branch, then every workflow is triggered.